requirements.txt
tensorflow>=2.13.0
customtkinter>=5.2.0
matplotlib>=3.7.0
numpy>=1.24.0
scipy>=1.11.0
Pillow>=10.0.0
